% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{set_keyfile}
\alias{set_keyfile}
\title{Set filename for authorization tokens.}
\usage{
set_keyfile(filename, service)
}
\arguments{
\item{filename}{Name of file storing keys.}

\item{service}{Name of service key to return (google, opencage, etc).}
}
\description{
Set filename for authorization tokens.
}
\details{
The authorization file should be of the form:

service authorization_key

Supported services are opencage and googlemaps.

So a line like

googlemaps IOUYGUBKJKBKNOlknlnlnlknLnkl

is what is required.

Blank lines and lines beginning with # are skipped
}
